make it compile on gcc 4.1 and 4.3, fix a severe bug in the decoupled models
the saturation argument for the initSomething class was an int instead
of a Scalar. probably nobody noticed so far because the saturation was
always initialized to 1.0 in the decoupled tests

git-svn-id: svn://svn.iws.uni-stuttgart.de/DUMUX/dumux/trunk@5148 2fb0f335-1f38-0410-981e-8018bf24f1b0

+++ appl/lecture/msm/mcwhorter/mcwhorter_ff.cc
@@ -40,7 +40,7 @@ int main(int argc, char** argv)
         //load interface-file
          Dumux::InterfaceProblemProperties interfaceProbProps("interface_MW.xml");
          double discretizationLength = interfaceProbProps.IPP_DiscretizationLength;
-         int cellNumber = 2/discretizationLength;
+         int cellNumber = static_cast<int>(2/discretizationLength);
 
         // define the problem dimensions
         Dune::FieldVector<Scalar, dim> lowerLeft(0);
